backend/requirements.txt
Flask
Flask-Cors
pytesseract
Pillow
requests
numpy
opencv-python-headless
imutils 